+++ example/gg_example.dart
@@ -5,11 +5,11 @@
 // Use of this source code is governed by terms that can be
 // found in the LICENSE file in the root of this package.
 
-import '../bin/kd.dart';
+import '../bin/gg.dart';
 
 Future<void> main() async {
   print('Executing');
-  await runKidney(
+  await runGg(
     args: ['bash', '--verbose', 'ls', '-l', '-a'],
     ggLog: (String message) {
       print(message);