+++ lib/src/commands/gg_one.dart
@@ -6,14 +6,14 @@
 // found in the LICENSE file in the root of this package.
 
 import 'package:args/command_runner.dart';
-import 'package:gg/gg.dart';
+import 'package:gg_one/gg_one.dart' as gg_one;
 import 'package:gg_log/gg_log.dart';
 
-/// Command that exposes all gg subcommands under the kidney namespace.
-class KidneyOne extends Command<void> {
+/// Command that exposes all gg subcommands under the gg namespace.
+class GgOne extends Command<void> {
   /// Create the command and register all gg subcommands.
-  KidneyOne({required this.ggLog}) {
-    Gg(ggLog: ggLog).subcommands.values.forEach(addSubcommand);
+  GgOne({required this.ggLog}) {
+    gg_one.Gg(ggLog: ggLog).subcommands.values.forEach(addSubcommand);
   }
 
   /// The log function.