+++ test/commands/gg_one_test.dart
@@ -4,21 +4,21 @@
 // Use of this source code is governed by terms that can be
 // found in the LICENSE file in the root of this package.
 
-import 'package:gg/gg.dart';
+import 'package:gg_one/gg_one.dart' as gg_one;
 import 'package:gg_log/gg_log.dart';
-import 'package:kd/src/commands/kidney_one.dart';
+import 'package:gg/src/commands/gg_one.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('KidneyOne', () {
-    late KidneyOne command;
+  group('GgOne', () {
+    late GgOne command;
     late List<String> messages;
     late GgLog ggLog;
 
     setUp(() {
       messages = <String>[];
       ggLog = messages.add;
-      command = KidneyOne(ggLog: ggLog);
+      command = GgOne(ggLog: ggLog);
     });
 
     test('returns the expected name', () {
@@ -30,14 +30,14 @@ void main() {
     });
 
     test('registers all gg subcommands', () {
-      final expectedSubcommands = Gg(ggLog: ggLog).subcommands;
+      final expectedSubcommands = gg_one.Gg(ggLog: ggLog).subcommands;
 
       expect(command.subcommands.keys, expectedSubcommands.keys);
       expect(command.subcommands, hasLength(expectedSubcommands.length));
     });
 
     test('exposes the same subcommand instances by name type contract', () {
-      final expectedSubcommands = Gg(ggLog: ggLog).subcommands;
+      final expectedSubcommands = gg_one.Gg(ggLog: ggLog).subcommands;
 
       for (final entry in expectedSubcommands.entries) {
         final actualSubcommand = command.subcommands[entry.key];